+++ spec/dsu/support/descriptable_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.describe Dsu::Support::Descriptable, type: :module do
+  subject(:descriptable) { described_class }
+
+  it 'does something awesome'
+end